touch up hostdb
